yask: Move to dedicated directory [finished]
